django_roomsmandu_backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
psycopg2-binary==2.9.11
dj-rest-auth==7.0.1
django-allauth==65.13.1
django-cors-headers==4.9.0
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
channels==4.3.2
daphne==4.2.1 #WEBSOCKET
requests==2.32.3
